etl/requirements.txt
psycopg2-binary
boto3
flask>=0.12
